+++ src/visor/db.go
@@ -154,11 +154,11 @@ func backupDB(db *dbutil.DB) (*dbutil.DB, error) { // nolint: unused,megacheck
 	return OpenDB(dbPath, dbReadOnly)
 }
 
-// RepairCorruptDB checks the database for corruption and if corrupted and
+// ResetCorruptDB checks the database for corruption and if corrupted and
 // is ErrMissingSignature, then then it erases the db and starts over.
 // If it's ErrHistoryDBCorrupted, then rebuild historydb from scratch.
 // A copy of the corrupted database is saved.
-func RepairCorruptDB(db *dbutil.DB, pubkey cipher.PubKey, quit chan struct{}) (*dbutil.DB, error) {
+func ResetCorruptDB(db *dbutil.DB, pubkey cipher.PubKey, quit chan struct{}) (*dbutil.DB, error) {
 	err := CheckDatabase(db, pubkey, quit)
 	switch err.(type) {
 	case nil: